+++ packages/backend-common/src/reading/GiteaUrlReader.ts
@@ -34,6 +34,7 @@ import {
   NotModifiedError,
 } from '@backstage/errors';
 import { Readable } from 'stream';
+import { parseLastModified } from './util';
 
 /**
  * Implements a {@link @backstage/backend-plugin-api#UrlReaderService} for the Gitea v1 api.
@@ -86,6 +87,9 @@ export class GiteaUrlReader implements UrlReader {
           Readable.from(Buffer.from(content, 'base64')),
           {
             etag: response.headers.get('ETag') ?? undefined,
+            lastModifiedAt: parseLastModified(
+              response.headers.get('Last-Modified'),
+            ),
           },
         );
       }