+++ plugins/entity-feedback/src/api/EntityFeedbackApi.ts
@@ -20,6 +20,7 @@ import {
   EntityRatingsData,
   FeedbackResponse,
   Rating,
+  Ratings,
 } from '@backstage/plugin-entity-feedback-common';
 
 /**
@@ -40,6 +41,11 @@ export interface EntityFeedbackApi {
   recordRating(entityRef: string, rating: string): Promise<void>;
 
   getRatings(entityRef: string): Promise<Omit<Rating, 'entityRef'>[]>;
+
+  /**
+   * Returns anonymized aggregated results for one entityRef
+   */
+  getRatingAggregates(entityRef: string): Promise<Ratings>;
 
   recordResponse(
     entityRef: string,