+++ plugins/scaffolder/src/alpha.ts
@@ -16,9 +16,6 @@
 
 export { NextScaffolderPage } from './plugin';
 export {
-  nextRouteRef,
-  nextScaffolderTaskRouteRef,
-  nextSelectedTemplateRouteRef,
   type TemplateGroupFilter,
   type NextRouterProps,
   type FormProps,