+++ plugins/scaffolder/src/next/TemplateWizardPage/TemplateWizardPage.tsx
@@ -34,11 +34,12 @@ import {
 } from '@backstage/plugin-scaffolder-react/alpha';
 import { JsonValue } from '@backstage/types';
 import { Header, Page } from '@backstage/core-components';
+
 import {
-  nextRouteRef,
-  nextScaffolderTaskRouteRef,
-  nextSelectedTemplateRouteRef,
-} from '../routes';
+  rootRouteRef,
+  scaffolderTaskRouteRef,
+  selectedTemplateRouteRef,
+} from '../../routes';
 
 export type TemplateWizardPageProps = {
   customFieldExtensions: NextFieldExtensionOptions<any, any>[];
@@ -47,13 +48,13 @@ export type TemplateWizardPageProps = {
 };
 
 export const TemplateWizardPage = (props: TemplateWizardPageProps) => {
-  const rootRef = useRouteRef(nextRouteRef);
-  const taskRoute = useRouteRef(nextScaffolderTaskRouteRef);
+  const rootRef = useRouteRef(rootRouteRef);
+  const taskRoute = useRouteRef(scaffolderTaskRouteRef);
   const { secrets } = useTemplateSecrets();
   const scaffolderApi = useApi(scaffolderApiRef);
   const navigate = useNavigate();
   const { templateName, namespace } = useRouteRefParams(
-    nextSelectedTemplateRouteRef,
+    selectedTemplateRouteRef,
   );
 
   const templateRef = stringifyEntityRef({