+++ packages/catalog-client/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @backstage/catalog-client
 
+## 1.4.0
+
+### Minor Changes
+
+- 8c030b1826c: Add `queryEntities` method to `CatalogApi`.
+
+### Patch Changes
+
+- 52b0022dab7: Updated dependency `msw` to `^1.0.0`.
+- c630360631f: Ensure that `getEntitiesByRefs` returns `undefined` instead of `null` for missing items
+- Updated dependencies
+  - @backstage/errors@1.1.5
+  - @backstage/catalog-model@1.2.1
+
 ## 1.4.0-next.1
 
 ### Patch Changes