+++ plugins/adr/CHANGELOG.md
@@ -1,5 +1,22 @@
 # @backstage/plugin-adr
 
+## 0.4.1
+
+### Patch Changes
+
+- d293f4aaf83: Updated readme instructions
+- 52b0022dab7: Updated dependency `msw` to `^1.0.0`.
+- Updated dependencies
+  - @backstage/core-components@0.12.5
+  - @backstage/plugin-catalog-react@1.4.0
+  - @backstage/plugin-search-react@1.5.1
+  - @backstage/core-plugin-api@1.5.0
+  - @backstage/catalog-model@1.2.1
+  - @backstage/integration-react@1.1.11
+  - @backstage/theme@0.2.18
+  - @backstage/plugin-adr-common@0.2.7
+  - @backstage/plugin-search-common@1.2.2
+
 ## 0.4.1-next.2
 
 ### Patch Changes