+++ plugins/cost-insights/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @backstage/plugin-cost-insights
 
+## 0.12.5
+
+### Patch Changes
+
+- 928a12a9b3e: Internal refactor of `/alpha` exports.
+- 52b0022dab7: Updated dependency `msw` to `^1.0.0`.
+- Updated dependencies
+  - @backstage/core-components@0.12.5
+  - @backstage/plugin-catalog-react@1.4.0
+  - @backstage/core-plugin-api@1.5.0
+  - @backstage/catalog-model@1.2.1
+  - @backstage/config@1.0.7
+  - @backstage/theme@0.2.18
+  - @backstage/plugin-cost-insights-common@0.1.1
+
 ## 0.12.5-next.2
 
 ### Patch Changes