+++ plugins/permission-react/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @backstage/plugin-permission-react
 
+## 0.4.11
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/core-plugin-api@1.5.0
+  - @backstage/plugin-permission-common@0.7.4
+  - @backstage/config@1.0.7
+
 ## 0.4.11-next.2
 
 ### Patch Changes