+++ plugins/splunk-on-call/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @backstage/plugin-splunk-on-call
 
+## 0.4.5
+
+### Patch Changes
+
+- 65454876fb2: Minor API report tweaks
+- 52b0022dab7: Updated dependency `msw` to `^1.0.0`.
+- Updated dependencies
+  - @backstage/core-components@0.12.5
+  - @backstage/plugin-catalog-react@1.4.0
+  - @backstage/core-plugin-api@1.5.0
+  - @backstage/catalog-model@1.2.1
+  - @backstage/theme@0.2.18
+
 ## 0.4.5-next.2
 
 ### Patch Changes